requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
djangorestframework==3.13.1
docopt==0.6.2
idna==3.3
pipreqs==0.4.11
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
yarg==0.1.9
pyyaml==5.3.1
django-rest-swagger==2.2.0
django-colorfield==0.6.3
django-ckeditor==6.2.0
stripe==2.68.0
django-cors-headers==3.5.0
python-dateutil==2.8.2
xhtml2pdf==0.2.7
pdfkit==1.0.0
SQLAlchemy==1.4.31
wkhtmltopdf==0.2
pisa==3.0.33
moviepy==1.0.3
django-rosetta==0.9.8 
opencv-python==4.5.5.64
django-environ==0.8.1
translate==3.6.1
django-crontab==0.7.1
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
django-debug-toolbar==3.5.0
